Add support for thread based MPMs
This adds a global mutex (which is not optimal, but for loads a
server handling dynamic content manageable) and a per-pool cache
for the resolved path. The document root is set per request
so no cleanup is needed anymore.

For sake of simplicity we decided not to got with double buffering
and removing the mutex was not worth it for now.